output/V-KYC 2/app/pyproject.toml
[tool.poetry]
name = "fastapi-elk-app"
version = "0.1.0"
description = "FastAPI application integrated with ELK for centralized logging."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-json-logger = "^0.1.11"
loguru = "^0.7.2" # For structured logging and easy integration
python-dotenv = "^1.0.1" # For environment variable management

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"
locust = "^2.22.0"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
bandit = "^1.7.9"
safety = "^3.2.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120